fix uni fetching

+++ src/main/java/pro/belbix/ethparser/web3/uniswap/db/UniswapDbService.java
@@ -66,13 +66,13 @@ public class UniswapDbService {
             return uniswapRepository.fetchAllFromBlockDate(
                 Instant.now().minus(1, DAYS).toEpochMilli() / 1000);
         }
-        int fromI = 0;
-        int toI = Integer.MAX_VALUE;
+        long fromI = 0;
+        long toI = Integer.MAX_VALUE;
         if (from != null) {
-            fromI = Integer.parseInt(from);
+            fromI = Long.parseLong(from);
         }
         if (to != null) {
-            toI = Integer.parseInt(to);
+            toI = Long.parseLong(to);
         }
         return uniswapRepository.fetchAllByPeriod(fromI, toI);
     }
